fix: test case for the pipeline and service changes
